pyproject.toml
[build-system]
requires = ['setuptools>=72']
build-backend = 'setuptools.build_meta'

[project]
name = "dask-felleskomponenter"
version = "0.2.1-a5"
authors = [
  { name="Dataplattform@Statens Kartverk", email="dataplattform@kartverket.no" },
]
description = "Felleskomponenter på DASK"
readme = "README.md"
requires-python = ">=3.10, <3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28",
    "databricks-connect>=15.0",
    "psycopg>=3.0",
    "pyspark>=4.1.1",
    "zstandard>=0.25.0",
    "datacontract-cli[databricks] (>=0.11.6,<0.12.0)",
    "typer (>=0.24.1,<0.25.0)",
    "semver (>=3.0.4,<4.0.0)",
]

[project.scripts]
validate-data-contract = "dask_felleskomponenter.governance.data_contract.validate:app"

[project.urls]
Homepage = "https://github.com/kartverket/dask-felleskomponenter"
Issues = "https://github.com/kartverket/dask-felleskomponenter/issues"

[tool.setuptools.packages.find]
where = ["src"]



[dependency-groups]
dev = [
    "black>=26.0.0",
    "coverage>=7.13.5",
]
